fixes play button
