+++ project/src/components/buttons/button-play.jsx
@@ -1,27 +1,29 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { useHistory } from 'react-router';
+// import React from 'react';
+// import PropTypes from 'prop-types';
+// import { useHistory } from 'react-router';
 
-function PlayButton(props) {
-  const {id} = props;
-  const history = useHistory();
+// function PlayButton(props) {
+//   const {id} = props;
+//   const history = useHistory();
+//   /* eslint-disable no-console */
+//   console.log(id);
 
-  return (
-    <button
-      className="btn btn--play film-card__button"
-      type="button"
-      onClick={() => history.push(`/player/${id}`)}
-    >
-      <svg viewBox="0 0 19 19" width="19" height="19">
-        <use xlinkHref="#play-s"></use>
-      </svg>
-      <span>Play</span>
-    </button>
-  );
-}
+//   return (
+//     <button
+//       className="btn btn--play film-card__button"
+//       type="button"
+//       onClick={() => history.push(`/player/${id}`)}
+//     >
+//       <svg viewBox="0 0 19 19" width="19" height="19">
+//         <use xlinkHref="#play-s"></use>
+//       </svg>
+//       <span>Play</span>
+//     </button>
+//   );
+// }
 
-PlayButton.propTypes = {
-  id: PropTypes.number.isRequired,
-};
+// PlayButton.propTypes = {
+//   id: PropTypes.number.isRequired,
+// };
 
-export default PlayButton;
+// export default PlayButton;
