+++ project/src/components/main-page/main-page.jsx
@@ -1,18 +1,21 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 import MovieList from '../movie-list/movie-list';
+import { useHistory } from 'react-router';
 import Logo from '../logo/logo';
 import Footer from '../footer/footer';
 import Genres from '../genre-list/genre-list';
 import UserBlock from '../user-block/user-block';
-import PlayButton from '../buttons/button-play';
+// import PlayButton from '../buttons/button-play';
 import MyListButton from '../buttons/button-my-list';
 import HiddenPart from './hidden-part';
 import { getPromoMovie } from '../../store/movie-data/selectors';
 
 function MainPage() {
   const promoMovie = useSelector(getPromoMovie);
-  const {name, genre, released, posterImage, backgroundImage} = promoMovie;
+  const {id, name, genre, released, posterImage, backgroundImage} = promoMovie;
+
+  const history = useHistory();
 
   return (
     <>
@@ -43,7 +46,17 @@ function MainPage() {
               </p>
 
               <div className="film-card__buttons">
-                <PlayButton id={promoMovie.id}/>
+                {/* <PlayButton id={id}/> */}
+                <button
+                  className="btn btn--play film-card__button"
+                  type="button"
+                  onClick={() => history.push(`/player/${id}`)}
+                >
+                  <svg viewBox="0 0 19 19" width="19" height="19">
+                    <use xlinkHref="#play-s"></use>
+                  </svg>
+                  <span>Play</span>
+                </button>
                 <MyListButton />
               </div>
             </div>
