+++ project/src/components/movie-page/movie-page.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 import { Link, useParams } from 'react-router-dom';
+import { useHistory } from 'react-router';
 import Logo from '../logo/logo';
 import UserBlock from '../user-block/user-block';
-import PlayButton from '../buttons/button-play';
 import MyListButton from '../buttons/button-my-list';
 import TabLinks from '../tabs/tabs';
 import SimiliarMovies from '../similiar-movies/similiar-movies';
@@ -16,6 +16,7 @@ function MoviePage() {
   const movies = useSelector(getMovies);
   const params = useParams();
   const authorizationStatus = useSelector(getAuthorizationStatus);
+  const history = useHistory();
 
   const [currentMovie] = movies.filter((item) => item.id === +params.id);
   const {id, name, previewImage, genre, released, backgroundImage} = currentMovie;
@@ -44,7 +45,17 @@ function MoviePage() {
               </p>
 
               <div className="film-card__buttons">
-                <PlayButton id={currentMovie.id}/>
+                {/* <PlayButton id={id}/> */}
+                <button
+                  className="btn btn--play film-card__button"
+                  type="button"
+                  onClick={() => history.push(`/player/${id}`)}
+                >
+                  <svg viewBox="0 0 19 19" width="19" height="19">
+                    <use xlinkHref="#play-s"></use>
+                  </svg>
+                  <span>Play</span>
+                </button>
                 <MyListButton />
                 {
                   authorizationStatus === AuthorizationStatus.AUTH &&
@@ -72,7 +83,9 @@ function MoviePage() {
         </div>
       </section>
       <div className="page-content">
-        <SimiliarMovies currentMovie={currentMovie} />
+        <SimiliarMovies
+          currentMovie={currentMovie}
+        />
         <Footer />
       </div>
     </>
